requirements.txt
wheel
selenium
requests
yt_dlp
python-dotenv
webdriver-manager
pre-commit
yapf
